+++ Interactive-rating-component/src/components/Button/Button.jsx
