+++ Interactive-rating-component/src/components/Button/Button.module.css
@@ -1,5 +1,5 @@
 /* Variable Imports */
-@import '../variables.css';
+@import '../../variables.css';
 
 .button {
   all: unset;
