+++ Interactive-rating-component/src/components/Image/Image.jsx
