+++ Interactive-rating-component/src/components/Response/Response.jsx
@@ -1,0 +1,8 @@
+/* Styling imports */
+import styles from './Rating.module.css'
+
+/* Component Imports */
+import Image from '../Image/Image';
+
+/* Image Imports */
+import ThankYou from '../../../assets/images/illustration-thank-you.svg'
