+++ Interactive-rating-component/src/components/Response/Response.module.css
@@ -1,0 +1,2 @@
+/* Variable Imports */
+@import '../../variables.css';
